requirements.txt
streamlit==1.0.0
plotly==5.3.1
pandas==1.1.3
Pillow==8.3.2
xlrd<2.0